Foreign.CUDA.FFT.Error

Description

Documentation

data Result

Error codes used by cuFFT library functions

http://docs.nvidia.com/cuda/cufft/index.html#cufftresult

Constructors

Success

InvalidPlan

AllocFailed

InvalidType

InvalidValue

InternalError

ExecFailed

SetupFailed

InvalidSize

UnalignedData

IncompleteParameterList

InvalidDevice

ParseError

NoWorkspace

NotImplemented

LicenseError

NotSupported

describe :: Result -> String

Describe an error code

data CUFFTException

Constructors

ExitCode Result

UserError String

cufftError :: String -> IO a

Raise a CUFFTException in the IO Monad

resultIfOk :: (Result, a) -> IO a

Return the results of a function on successful execution, otherwise throw an exception with an error string associated with the return code

nothingIfOk :: Result -> IO ()

Throw an exception with an error string associated with an unsuccessful return code, otherwise return unit.

checkStatus :: CInt -> IO ()