Foreign.CUDA.FFT.Execute

Description

Documentation

data Mode

FFT transform direction

Constructors

Forward

Inverse

execC2C

Arguments

:: Handle
plan handle, of type C2C

-> Mode
transform direction

-> DevicePtr (Complex Float)
input data

-> DevicePtr (Complex Float)
output data

-> IO ()

Executes a single-precision complex-to-complex transform.

If the input and output device pointers are the same, an in-place transform is executed.

http://docs.nvidia.com/cuda/cufft/index.html#function-cufftexecc2c-cufftexecz2z

execZ2Z

Arguments

:: Handle
plan handle, of type Z2Z

-> Mode
transform direction

-> DevicePtr (Complex Double)
input data

-> DevicePtr (Complex Double)
output data

-> IO ()

Executes a double-precision complex-to-complex transform.

If the input and output device pointers are the same, an in-place transform is executed.

http://docs.nvidia.com/cuda/cufft/index.html#function-cufftexecc2c-cufftexecz2z

execR2C

Arguments

:: Handle
plan handle, of type R2C

-> DevicePtr Float
input data

-> DevicePtr (Complex Float)
output data

-> IO ()

Executes a single-precision real-to-complex, implicitly forward, transform.

If the input and output device pointers refer to the same address, an in-place transform is executed.

http://docs.nvidia.com/cuda/cufft/index.html#function-cufftexecr2c-cufftexecd2z

execD2Z

Arguments

:: Handle
plan handle, of type D2Z

-> DevicePtr Double
input data

-> DevicePtr (Complex Double)
output data

-> IO ()

Executes a double-precision real-to-complex, implicitly forward, transform.

If the input and output device pointers refer to the same address, an in-place transform is executed.

http://docs.nvidia.com/cuda/cufft/index.html#function-cufftexecr2c-cufftexecd2z

execC2R

Arguments

:: Handle
plan handle, of type C2R

-> DevicePtr (Complex Float)
input data

-> DevicePtr Float
output data

-> IO ()

Executes a single-precision complex-to-real, implicitly forward, transform.

If the input and output device pointers refer to the same address, an in-place transform is executed.

http://docs.nvidia.com/cuda/cufft/index.html#function-cufftexecc2r-cufftexecz2d

execZ2D

Arguments

:: Handle
plan handle, of type Z2D

-> DevicePtr (Complex Double)
input data

-> DevicePtr Double
output data

-> IO ()

Executes a double-precision complex-to-real, implicitly forward, transform.

If the input and output device pointers refer to the same address, an in-place transform is executed.

http://docs.nvidia.com/cuda/cufft/index.html#function-cufftexecc2r-cufftexecz2d