Foreign.CUDA.FFT.Stream

Description

Documentation

data Stream :: *

A processing stream. All operations in a stream are synchronous and executed in sequence, but operations in different non-default streams may happen out-of-order or concurrently with one another.

Use Events to synchronise operations between streams.

setStream :: Handle -> Stream -> IO ()

Set the execution stream which all subsequent cuFFT library functions will execute with. This enables the activity in this execution stream (e.g. kernel launches and data transfer) to overlap with activity in other execution streams. The association remains until the plan is destroyed or the stream is changed.

If not set, functions execute in the default stream, which never overlaps with any other operation.

http://docs.nvidia.com/cuda/cufft/index.html#function-cufftsetstream